calculate log
